move function to test context
