+++ test/support/tenant_connection.ex
@@ -2,6 +2,9 @@ defmodule TenantConnection do
   @moduledoc """
   Boilerplate code to handle Realtime.Tenants.Connect during tests
   """
+  alias Realtime.Api.Message
+  alias Realtime.Database
+  alias Realtime.Repo
 
   def broadcast_test_message(conn, private, topic, event, payload) do
     query =
@@ -18,5 +21,22 @@ defmodule TenantConnection do
       """
 
     Postgrex.query!(conn, query, [private, topic, event, %{payload: payload}])
+  end
+
+  def create_message(attrs, conn, opts \\ [mode: :savepoint]) do
+    channel = Message.changeset(%Message{}, attrs)
+
+    {:ok, result} =
+      Database.transaction(conn, fn transaction_conn ->
+        with {:ok, %Message{} = channel} <- Repo.insert(transaction_conn, channel, Message, opts) do
+          channel
+        end
+      end)
+
+    case result do
+      %Ecto.Changeset{valid?: false} = error -> {:error, error}
+      {:error, error} -> {:error, error}
+      result -> {:ok, result}
+    end
   end
 end
